change prefix of coin specific actions
